requirements.txt
mysql-connector-python
pandas
python-dotenv
pathlib
PyYAML
pylint
black
pytest
boto3
fsspec
s3fs
gspread
oauth2client
google-api-python-client